feat(agent): Commandcode agent setup. Clippy fix.
