+++ rust/src/core/editor_registry/writers/install/copilot.rs
@@ -1,0 +1,81 @@
+#[allow(clippy::wildcard_imports)]
+use super::super::shared::*;
+use super::super::{WriteAction, WriteOptions, WriteResult};
+use crate::core::editor_registry::types::EditorTarget;
+
+pub(crate) fn write_copilot_cli(
+    target: &EditorTarget,
+    binary: &str,
+    opts: WriteOptions,
+) -> Result<WriteResult, String> {
+    let desired = serde_json::json!({
+        "type": "local",
+        "command": binary,
+        "args": ["mcp"],
+        "tools": ["*"]
+    });
+
+    if target.config_path.exists() {
+        let content = std::fs::read_to_string(&target.config_path).map_err(|e| e.to_string())?;
+        let mut json = match crate::core::jsonc::parse_jsonc(&content) {
+            Ok(v) => v,
+            Err(_e) => {
+                return handle_invalid_json_write(
+                    &target.config_path,
+                    &content,
+                    "mcpServers",
+                    "lean-ctx",
+                    &desired,
+                    opts.overwrite_invalid,
+                );
+            }
+        };
+        let obj = json
+            .as_object_mut()
+            .ok_or_else(|| "root JSON must be an object".to_string())?;
+
+        let servers = obj
+            .entry("mcpServers")
+            .or_insert_with(|| serde_json::json!({}));
+        let servers_obj = servers
+            .as_object_mut()
+            .ok_or_else(|| "\"mcpServers\" must be an object".to_string())?;
+
+        let existing = servers_obj.get("lean-ctx").cloned();
+        if existing.as_ref() == Some(&desired) {
+            return Ok(WriteResult {
+                action: WriteAction::Already,
+                note: None,
+            });
+        }
+
+        servers_obj.insert("lean-ctx".to_string(), desired);
+        let out = serde_json::to_string_pretty(&json).map_err(|e| e.to_string())?;
+        crate::config_io::write_atomic_with_backup(&target.config_path, &out)?;
+        return Ok(WriteResult {
+            action: WriteAction::Updated,
+            note: None,
+        });
+    }
+
+    // Fresh write
+    if let Some(parent) = target.config_path.parent() {
+        std::fs::create_dir_all(parent).map_err(|e| e.to_string())?;
+    }
+    let content = serde_json::to_string_pretty(&serde_json::json!({
+        "mcpServers": {
+            "lean-ctx": {
+                "type": "local",
+                "command": binary,
+                "args": ["mcp"],
+                "tools": ["*"]
+            }
+        }
+    }))
+    .map_err(|e| e.to_string())?;
+    crate::config_io::write_atomic_with_backup(&target.config_path, &content)?;
+    Ok(WriteResult {
+        action: WriteAction::Created,
+        note: None,
+    })
+}
